+++ src/Cru/Common.h
@@ -71,6 +71,8 @@ struct Link {
   uint32_t feeId = 0x0;
   /*uint32_t systemId = 0x1ff; // invalid, used as "unset" placeholder
   uint32_t feeId = 0x1f; // invalid, used as "unset" placeholder*/
+  uint32_t glitchCounter = 0;
+  uint32_t fecCounter = 0;
 
   bool operator==(const Link& dlink) const
   {
@@ -115,6 +117,7 @@ struct ReportInfo {
   bool runStatsEnabled;
   bool userAndCommonLogicEnabled;
   uint16_t timeFrameLength;
+  bool dmaStatus;
 };
 
 struct OnuStickyStatus {