+++ src/FirmwareChecker.cxx
@@ -30,6 +30,8 @@ FirmwareChecker::FirmwareChecker() : mCompatibleFirmwareList({
                                        { "2058c933", "v3.14.0" },
                                        { "4a412c71", "v3.13.0" },
                                        { "82b4662d", "MFT PSU" },
+                                       { "6838510f", "v3.17.0" },
+                                       { "8e74a7f8", "v3.17.1" },
                                        /*{ "6a85d30c", "v3.12.0" },
                                        { "7be5aa1c", "v3.11.0" },
                                        { "e4a5a46e", "v3.10.0" },
@@ -45,6 +47,8 @@ FirmwareChecker::FirmwareChecker() : mCompatibleFirmwareList({
                                        { "267f8e5", "v2.9.1" },
                                        { "cecc295", "v2.9.0" },
                                        { "221ff280", "v2.10.0" },
+                                       { "cfa0bc9c", "2.10.1" },
+                                       { "2d4c9028", "2.11.0" },
                                        /*{ "59e9955", "v2.8.1" },
                                        { "f086417", "v2.8.0" },
                                        { "474f9e1", "v2.7.0" }